dexter/requirements.txt
# Dexter — Phase 1 deps. Pin versions in production.
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
